Rename `use-active-link`
